Functionality complete, stats remain.

+++ include/process.h
@@ -24,7 +24,7 @@ private:
     int32_t cpu_time;           // total time spent running on a CPU core
     int32_t remain_time;        // CPU time remaining until terminated
     uint64_t launch_time;       // actual time in ms (since epoch) that process was 'launched'
-    uint64_t last_update_time;  // time since last call to updateProcess()
+    uint32_t last_update_time;  // time since last call to updateProcess()
     // you are welcome to add other private data fields here if you so choose
 
 public:
